third_party/blink/web_tests/external/wpt/css/css-conditional/js/CSS-supports-CSSStyleDeclaration-expected.txt

This is a testharness.js-based test.
Found 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] -epub-caption-side: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-text-combine: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-text-emphasis: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-text-emphasis-color: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-text-emphasis-style: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-text-orientation: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-text-transform: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-word-break: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] -epub-writing-mode: _camel_cased_attribute v. CSS.supports
  assert_equals: expected false but got true
[FAIL] font-display: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] font-display: _dashed_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] inherits: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] initial-value: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] initial-value: _dashed_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] speak-as: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] speak-as: _dashed_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] src: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] syntax: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] unicode-range: _camel_cased_attribute v. CSS.supports
  assert_equals: expected true but got false
[FAIL] unicode-range: _dashed_attribute v. CSS.supports
  assert_equals: expected true but got false
Harness: the test ran to completion.
